revise quizzes.js DELETE route for quizzes/:quizzes_id/respondents/:respondent_id

+++ backend/routes/quizzes.js
@@ -709,13 +709,33 @@ router.delete(
           message: "Unauthorized!",
         });
       }
-      request.userId = decoded.id;
       pool.query(
-        "DELETE FROM respondents WHERE quizzes_id=$1 AND id=$2",
-        [quizzes_id, respondent_id],
+        "SELECT * FROM quizzes WHERE id=$1",
+        [quizzes_id],
         (err, res) => {
           if (err) return next(err);
-          response.status(204).end();
+          if (res.rows.length === 0)
+            return response.json({
+              error: true,
+              message: "This quiz does not exist",
+            });
+          const user_id = res.rows[0].creators_id;
+          request.userId = decoded.id;
+          if (request.userId !== Number(user_id)) {
+            return response.json({
+              error: true,
+              message: "Token ID provided does not match!",
+            });
+          } else {
+            pool.query(
+              "DELETE FROM respondents WHERE quizzes_id=$1 AND id=$2",
+              [quizzes_id, respondent_id],
+              (err, res) => {
+                if (err) return next(err);
+                response.status(204).end();
+              }
+            );
+          }
         }
       );
     });
